+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file.
 - Created git-cliff action (#8)
 - Creat wlan bluetooth service (#10)
 - *(gps)* Create recorder package (#13)
+- Created logger package (#15)
 
 ### 🐛 Bug Fixes